feat(grammar): add Grammar
