+++ package.json
@@ -9,10 +9,10 @@
     "lint": "eslint"
   },
   "dependencies": {
-    "class-variance-authority": "^0.7.1",
-    "lucide-react": "^0.541.0",
     "@hookform/resolvers": "^5.2.1",
     "@reduxjs/toolkit": "^2.8.2",
+    "class-variance-authority": "^0.7.1",
+    "lucide-react": "^0.541.0",
     "next": "15.5.0",
     "next-auth": "^4.24.11",
     "openai": "^5.15.0",
